Refactor: extract the functions for querying the CodeScene REST API
